Move ams assignment from physical to telemissile action
